Add files via upload
A second go around at understanding the edatec code, how it works, what it does, and what I need for it to do. 
basically, all you need is the post install script from the deb file. the loading script is where it tries to set up all the crap
I dont want or need.